modify 'last post in conversation' logic
1. Instead of checking to see whether a post is the last in the
conversation, check to see whether it has any replies. Otherwise, if I
respond to a post, and then someone else responds to the same post, I
can no longer delete my status.

2. Move all the logic about whether a user can delete a given post to
the same place. This a) simplifies the view logic, and b) allows us
to skip checking every status's `in-reply-to` field if it's unnecessary
(because if the user isn't the author, it can't be deleted anyway).