fix #56  Entry form doesn't allow more than 140 chars
The solution allows to put more than 140 characters in the input field. The counter will count down to 0 and when it switches to negative values, the submit button is disabled so the input cannot be submitted. At the same time the "ENTER" hotkey on the input field is automatically disabled so the form can neither be submitted by pressing "ENTER".

The submit functionality is enabled again when the counter is back to 0 chars ore more.