Reply form shouldn't be redrawn for same post
If a reply form is visible and the user clicks the reply button on the same post again, the action should simply be ignored. Otherwise the form is removed, retrieved and displayed again which not only creates visual hassle but also (small) unnecessary traffic on the data plan to retrieve the reply form via HTTP call.